# Parallelize array deinitialization