--- a/test/constrained-generics/hashtable/MyHashtable.chpl
+++ b/test/constrained-generics/hashtable/MyHashtable.chpl
@@ -283,8 +283,8 @@ module MyHashtable {
       // Go through the full slots in the current table and run
       // chpl__autoDestroy on the index
       if _typeNeedsDeinit(keyType) || _typeNeedsDeinit(valType) {
-        if _deinitElementsIsParallel(keyType) &&
-           _deinitElementsIsParallel(valType) {
+        if _deinitElementsIsParallel(keyType, tableSize) &&
+           _deinitElementsIsParallel(valType, tableSize) {
           forall slot in _allSlots(tableSize) {
             ref aSlot = table[slot];
             if _isSlotFull(aSlot) {